refactor: split trust.py into modular package (#180)
Split the 2529-line trust.py into focused submodules:

- edges.py: DecayModel, TrustEdge, TrustEdge4D (462 lines)
- computation.py: compute_delegated_trust, compute_transitive_trust (196 lines)
- graph_store.py: TrustGraphStore, get_trust_graph_store (376 lines)
- service.py: TrustService and convenience functions (582 lines)
- federation.py: FederationTrustEdge, FederationMembershipRegistry (667 lines)
- __init__.py: Re-exports all public API for backward compatibility

All existing imports (from valence.privacy.trust import X) continue to work.
Tests pass locally.
